+++ source/Energinet.DataHub.MarketParticipant.ApplyDBMigrationsApp/Scripts/LocalDB/202303141130 Remove PermissionEicFunction and Permission.sql
@@ -1,0 +1,10 @@
+ALTER TABLE [dbo].[PermissionAuditLogEntry]
+DROP CONSTRAINT [FK_PermissionAuditLogEntry_PermissionId_Permission]
+GO
+
+ALTER TABLE [dbo].[Permission]
+DROP COLUMN [Created]
+GO
+
+DROP Table [dbo].[PermissionEicFunction]
+GO