+++ source/Energinet.DataHub.MarketParticipant.EntryPoint.WebApi/Controllers/ActorController.cs
@@ -15,11 +15,11 @@
 using System;
 using System.Threading.Tasks;
 using Energinet.DataHub.Core.App.Common.Abstractions.Users;
-using Energinet.DataHub.Core.App.Common.Security;
-using Energinet.DataHub.Core.App.WebApp.Authorization;
 using Energinet.DataHub.MarketParticipant.Application.Commands.Actor;
 using Energinet.DataHub.MarketParticipant.Application.Security;
+using Energinet.DataHub.MarketParticipant.Domain.Model.Permissions;
 using Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Extensions;
+using Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Security;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
@@ -62,7 +62,7 @@ namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Controllers
         }
 
         [HttpPost]
-        [AuthorizeUser(Permission.ActorManage)]
+        [AuthorizeUser(PermissionId.ActorManage)]
         public async Task<IActionResult> CreateActorAsync(CreateActorDto actorDto)
         {
             return await this.ProcessAsync(
@@ -83,7 +83,7 @@ namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Controllers
         }
 
         [HttpPut("{actorId:guid}")]
-        [AuthorizeUser(Permission.ActorManage)]
+        [AuthorizeUser(PermissionId.ActorManage)]
         public async Task<IActionResult> UpdateActorAsync(Guid actorId, ChangeActorDto changeActor)
         {
             return await this.ProcessAsync(