+++ source/Energinet.DataHub.MarketParticipant.EntryPoint.WebApi/Controllers/GridAreaController.cs
@@ -14,10 +14,10 @@
 
 using System;
 using System.Threading.Tasks;
-using Energinet.DataHub.Core.App.Common.Security;
-using Energinet.DataHub.Core.App.WebApp.Authorization;
 using Energinet.DataHub.MarketParticipant.Application.Commands.GridArea;
+using Energinet.DataHub.MarketParticipant.Domain.Model.Permissions;
 using Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Extensions;
+using Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Security;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
@@ -38,7 +38,7 @@ namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Controllers
         }
 
         [HttpPost]
-        [AuthorizeUser(Permission.GridAreasManage)]
+        [AuthorizeUser(PermissionId.GridAreasManage)]
         public async Task<IActionResult> CreateGridAreaAsync(CreateGridAreaDto gridAreaDto)
         {
             return await this.ProcessAsync(
@@ -56,7 +56,7 @@ namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Controllers
         }
 
         [HttpPut]
-        [AuthorizeUser(Permission.GridAreasManage)]
+        [AuthorizeUser(PermissionId.GridAreasManage)]
         public async Task<IActionResult> UpdateGridAreaAsync(ChangeGridAreaDto gridAreaDto)
         {
             return await this.ProcessAsync(
@@ -100,7 +100,7 @@ namespace Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Controllers
         }
 
         [HttpGet("{gridAreaId:guid}/auditlogentry")]
-        [AuthorizeUser(Permission.GridAreasManage)]
+        [AuthorizeUser(PermissionId.GridAreasManage)]
         public async Task<IActionResult> GetGridAreaAuditLogEntriesAsync(Guid gridAreaId)
         {
             return await this.ProcessAsync(