+++ source/Energinet.DataHub.MarketParticipant.EntryPoint.WebApi/Controllers/UserRoleController.cs
@@ -15,12 +15,12 @@
 using System;
 using System.Threading.Tasks;
 using Energinet.DataHub.Core.App.Common.Abstractions.Users;
-using Energinet.DataHub.Core.App.Common.Security;
-using Energinet.DataHub.Core.App.WebApp.Authorization;
 using Energinet.DataHub.MarketParticipant.Application.Commands.UserRoles;
 using Energinet.DataHub.MarketParticipant.Application.Security;
 using Energinet.DataHub.MarketParticipant.Domain.Model;
+using Energinet.DataHub.MarketParticipant.Domain.Model.Permissions;
 using Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Extensions;
+using Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Security;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
@@ -46,7 +46,7 @@ public sealed class UserRoleController : ControllerBase
     }
 
     [HttpGet]
-    [AuthorizeUser(Permission.UsersManage)]
+    [AuthorizeUser(PermissionId.UsersManage)]
     public async Task<IActionResult> GetAsync()
     {
         return await this.ProcessAsync(
@@ -64,7 +64,7 @@ public sealed class UserRoleController : ControllerBase
     }
 
     [HttpGet("{userRoleId:guid}")]
-    [AuthorizeUser(Permission.UsersManage)]
+    [AuthorizeUser(PermissionId.UsersManage)]
     public async Task<IActionResult> GetAsync(Guid userRoleId)
     {
         return await this.ProcessAsync(
@@ -82,7 +82,7 @@ public sealed class UserRoleController : ControllerBase
     }
 
     [HttpPost]
-    [AuthorizeUser(Permission.UserRoleManage)]
+    [AuthorizeUser(PermissionId.UserRoleManage)]
     public async Task<IActionResult> CreateAsync(CreateUserRoleDto userRole)
     {
         return await this.ProcessAsync(
@@ -100,7 +100,7 @@ public sealed class UserRoleController : ControllerBase
     }
 
     [HttpPut("{userRoleId:guid}")]
-    [AuthorizeUser(Permission.UserRoleManage)]
+    [AuthorizeUser(PermissionId.UserRoleManage)]
     public async Task<IActionResult> UpdateAsync(
         Guid userRoleId,
         UpdateUserRoleDto userRole)
@@ -121,7 +121,7 @@ public sealed class UserRoleController : ControllerBase
     }
 
     [HttpGet("{userRoleId:guid}/auditlogentry")]
-    [AuthorizeUser(Permission.UsersManage)]
+    [AuthorizeUser(PermissionId.UsersManage)]
     public async Task<IActionResult> GetUserRoleAuditLogsAsync(Guid userRoleId)
     {
         return await this.ProcessAsync(
@@ -139,7 +139,7 @@ public sealed class UserRoleController : ControllerBase
     }
 
     [HttpGet("permissions")]
-    [AuthorizeUser(Permission.UsersManage)]
+    [AuthorizeUser(PermissionId.UsersManage)]
     public async Task<IActionResult> GetPermissionDetailsAsync(EicFunction eicFunction)
     {
         return await this.ProcessAsync(
@@ -157,7 +157,7 @@ public sealed class UserRoleController : ControllerBase
     }
 
     [HttpGet("assignedtopermission")]
-    [AuthorizeUser(Permission.UsersManage)]
+    [AuthorizeUser(PermissionId.UsersManage)]
     public async Task<IActionResult> AssignedToPermissionAsync(int permissionId)
     {
         return await this.ProcessAsync(