+++ source/Energinet.DataHub.MarketParticipant.Infrastructure/Persistence/Repositories/UserRoleRepository.cs
@@ -16,8 +16,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Energinet.DataHub.Core.App.Common.Security;
 using Energinet.DataHub.MarketParticipant.Domain.Model;
+using Energinet.DataHub.MarketParticipant.Domain.Model.Permissions;
 using Energinet.DataHub.MarketParticipant.Domain.Model.Users;
 using Energinet.DataHub.MarketParticipant.Domain.Repositories;
 using Energinet.DataHub.MarketParticipant.Infrastructure.Persistence.Model;
@@ -78,18 +78,20 @@ public sealed class UserRoleRepository : IUserRoleRepository
     public async Task<UserRoleId> AddAsync(UserRole userRole)
     {
         ArgumentNullException.ThrowIfNull(userRole);
-        var role = new UserRoleEntity()
+
+        var role = new UserRoleEntity
         {
             Name = userRole.Name,
             Description = userRole.Description,
             Status = userRole.Status,
         };
-        foreach (var permissionEntity in userRole.Permissions.Select(x => new UserRolePermissionEntity() { Permission = x }))
+
+        foreach (var permissionEntity in userRole.Permissions.Select(x => new UserRolePermissionEntity { Permission = x }))
         {
             role.Permissions.Add(permissionEntity);
         }
 
-        role.EicFunctions.Add(new UserRoleEicFunctionEntity() { EicFunction = userRole.EicFunction });
+        role.EicFunctions.Add(new UserRoleEicFunctionEntity { EicFunction = userRole.EicFunction });
         _marketParticipantDbContext.UserRoles.Add(role);
         await _marketParticipantDbContext.SaveChangesAsync().ConfigureAwait(false);
         return new UserRoleId(role.Id);
@@ -124,7 +126,7 @@ public sealed class UserRoleRepository : IUserRoleRepository
         }
     }
 
-    public async Task<IEnumerable<UserRole>> GetAsync(Permission permission)
+    public async Task<IEnumerable<UserRole>> GetAsync(PermissionId permission)
     {
         var userRoles = await BuildUserRoleQuery()
             .Where(t => t